requirements.txt
django==1.9
django-tinymce==2.6.0
